grid practice
Learned about Grid Layout, which allows for proper alignment
however it's not as flexible as Flexbox. It allows for easy
structure and positioning.

You can change the layout with display: grid;
grids have either columns or rows for their templates,
you can change the layout with grid-template-columns,
grid-template-columns: px , fr;
each value after the property allows for accurate pixel space 'px'
or take up the rest of the space called freespace 'fr'.
Their is no limit to the amount of columns.