grid gaps
Learned how to create gaps between grids using the property;
column-gap and row-gap. Combined with the value freespace
it allows the screen to adjust based on the available space.

Still it's not as flexible as flexbox with the property like wrap
which allows the containers to wrap down or up based on the screen
availability.