CSS Positions
Learned about the positional property in CSS
Positon has 5 values, which I only learned about fixed and
static 'static is the intial value'.
Along with Position, you can affect the pixel value of how it is
position on the webpage with top,left,bottom,and right.
If you choose a positional pixel value in relation to the page
and also choose it's opposite side. i.e, top/bottom,left/right
you the property will stretch to acccomodate both/all sides.

Last key note, is that if you if you want the property to re-size
along with the page you can use the positional pixel values,
other-wise you want adjust with height or width so that it is
static.