Position Absolute
Learned a new value of position called Absolute, this places
the element on specifically on the page and scrolls with the page.

Position:fixed; is placed on the browser window and leaves the
current page contents behind when you scroll horizontally or
vertically.

Position: absolute; is placed on the webpage and is left behind
when the current page is scrolled horizontally or vertically from
the current browser window.

The absolute values are still the same and follows the same rules.
i.e, left,right,top,bottom and height/width.