z-index
Generally when div's or elements are written below another
element, they will appear infront of the layout. If a positional
fixed element is written before a positional absolute element
the absolute element will appear on top when you scroll the
browser window. To fix this the z-index property which has a
number value, tells the file the order. The z-index with the
highest order will appear on top of lower z-index values.