Position value Relative
I learned another value of Positon which is relative,
when an element has a positional css property: it behaves
normally with a keyfeature that allows other elements to be
in relation to it.

i.e, when an absolute positional element is placed inside
a relative positional element, the absolute element will
behave normally in the page when affected by the browser window.
It will not scroll along with the browser window.

This allows for great accuracy when positioning elements within
other elements that you want to behave normally.